+++ src/WrmPlugin.js
@@ -16,7 +16,7 @@ const {
 } = require('./helpers/web-resource-generator');
 const { toMap, extractPathPrefixForXml } = require('./helpers/options-parser');
 const { buildProvidedDependency } = require('./helpers/provided-dependencies');
-const { addBaseContext } = require('./settings/base-contexts');
+const { addBaseDependency } = require('./settings/base-dependencies');
 
 const ProvidedExternalDependencyModule = require('./webpack-modules/ProvidedExternalDependencyModule');
 const WrmDependencyModule = require('./webpack-modules/WrmDependencyModule');
@@ -212,7 +212,7 @@ An automated jsonpFunction name for this plugin was created:
                 'OverwritePublicPath Require-Extensions',
                 standardScript => {
                     // Ensure the `AJS.contextPath` function is available at runtime.
-                    addBaseContext('com.atlassian.plugins.atlassian-plugins-webresource-plugin:context-path');
+                    addBaseDependency('com.atlassian.plugins.atlassian-plugins-webresource-plugin:context-path');
 
                     // Add the public path extension to the webpack module runtime.
                     return `${standardScript}